The service once again changes the current working directory to the directory of the application. This change must have gotten lost somewhere when implementing the new config logic.
The service now exits after a pause request. This breaks windows guidelines (since it does not actually pause, but does a clean exit) This was done to implement the clean exit command in this way.

The service now notifies windows it can pause and continue.

In the rare event continue ever gets called it performs the start logic.